feat(tests): restructure tests
